+++ src/app/_components/navigation/locale_switcher/LocaleSwitcherDropdown.tsx
