+++ src/app/_components/navigation/navbar/logo/NavLogo.tsx
@@ -1,0 +1,10 @@
+import Image from "next/image";
+import Link from "next/link";
+
+export default function NavLogo() {
+    return (
+        <Link href="/">
+            <Image className="object-cover object-center" src="/logo.png" fill={true} alt="Logo" />
+        </Link>
+    )
+}
